+++ client/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useSessionBootstrap } from "./use-session-bootstrap.js";
 export { useRunLock } from "./use-run-lock.js";
 export { useMutations } from "./use-mutations.js";
+export { useCameraCommand } from "./use-camera-command.js";