+++ client/src/lib/index.ts
@@ -23,5 +23,8 @@ export {
   type MutationController,
   type MutationStore,
 } from "./mutations.js";
-export { createDebouncer, createCoalescingFlusher } from "./debounce.js";
-export { mutationErrors, type MutationError, type MutationErrorKey } from "./mutation-errors.js";
+export {
+  shouldApplyCameraCommand,
+  shouldConsumeCameraCommandSeq,
+} from "./camera-command.js";
+export { AGENT_MOVED_INDICATOR_MS, GRAPH_CAMERA_ANIMATION_MS } from "./camera-constants.js";